move ForwardMode into forward_batch
