+++ python/sglang/srt/layers/logits_processor.py
@@ -25,7 +25,8 @@ from vllm.distributed import (
     tensor_model_parallel_all_gather,
 )
 
-from sglang.srt.model_executor.model_runner import ForwardMode, InputMetadata
+from sglang.srt.model_executor.forward_batch import ForwardMode
+from sglang.srt.model_executor.model_runner import InputMetadata
 
 
 @dataclasses.dataclass
